Move monitoring wrapper code into its own module
This is to avoid capturing so much in the closure of the monitoring wrapper,
which prior to this patch, captured a lot of surprising stuff.

This reduces serialisation load (time and space).

With parsl/tests/test_bash/test_basic.py, using Work Queue and monitoring,
the five serialized function sizes are:

before this commit: 8400 bytes
after this commit: 7252 bytes

saving 1148 bytes per function invocation.
